refactor: move non-types from ast.ts
Fixes #1402
