+++ src/ast/ast-printer.spec.ts
@@ -5,7 +5,7 @@ import { join } from "path";
 import { trimTrailingCR, CONTRACTS_DIR } from "../test/util";
 import * as assert from "assert";
 import JSONBig from "json-bigint";
-import { getAstFactory } from "./ast";
+import { getAstFactory } from "./ast-helpers";
 import { defaultParser } from "../grammar/grammar";
 
 describe("formatter", () => {
