+++ src/ast/clone.ts
@@ -1,4 +1,5 @@
-import { AstNode, FactoryAst } from "./ast";
+import { AstNode } from "./ast";
+import { FactoryAst } from "./ast-helpers";
 import { throwInternalCompilerError } from "../error/errors";
 
 export function cloneNode<T extends AstNode>(
